AbstractClient: fix sendToOutput edge case
There are situations where a window can be completely outside of the old screen
area, for example when a XdgToplevelClient gets first initialized. When this
happens sendToOutput will not put the new window on the specified output. To
prevent that from happening, explicitly put the window inside the new screen
area.